source code of js, css and html
The webapp is made of html, css and js. It uses browsers speechsynthesis to convert text into speech and speech into text in real time. The voice may vary depending on the browser you are using. And it supports all kinds of languages, by default, it detects the language of your browser.